slider complete
